Translates text to a specific language

Use a Large Language Model (LLM) to translate a text to a specific language

Usage

llm_translate(
  .data,
  col,
  language,
  pred_name = ".translation",
  additional_prompt = ""
)

llm_vec_translate(x, language, additional_prompt = "")

Arguments

.data

A data.frame or tbl object that contains the text to be analyzed

col

The name of the field to analyze, supports tidy-eval

language

Target language to translate the text to

pred_name

A character vector with the name of the new column where the prediction will be placed

additional_prompt

Inserts this text into the prompt sent to the LLM

x

A vector that contains the text to be analyzed

Value

llm_translate returns a data.frame or tbl object. llm_vec_translate returns a vector that is the same length as x.